revert some changes, clean up
