+++ src/systemTest/resources/system-test-hub.properties
@@ -1,7 +1,7 @@
 helm.chart.path=http://ddt-nexus.flightstats.io/repository/helm/hub.tgz
-helm.release.name=ddt-hessp-5ojt
+# helm.release.name=
 # default is true
-helm.release.delete=false
+# helm.release.delete=false
 # default is true
 # helm.release.hub.isClustered=false
